# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.3
info:
  title: workspace API
  version: 1.0-SNAPSHOT
servers:
  - url: /workstation
paths:
  /racks:
    get:
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/Status'    
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rack'    
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rack'    
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rack'    
  /racks/{id}:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Rack ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rack'    
        "404":
          description: Not Found
    put:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Rack ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rack'    
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rack'    
        "404":
          description: Not Found
    delete:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Rack ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rack'    
        "404":
          description: Not Found

components:
  schemas:
    Status:
      type: string
      nullable: true
      enum:
        - ACTIVE
        - INACTIVE
        - null
    Rack:
      type: object
      properties:
        id:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        serialNumber:
          type: string
        status:
          $ref: '#/components/schemas/Status'    